requirements.txt
faiss-cpu
sentence-transformers
pypdf
langchain
langchain-community
python-dotenv
ollama
prometheus-client==0.17.1
psutil==5.9.5
flask-cors